+++ config/splat.us.stnz0.yaml
@@ -80,12 +80,14 @@ segments:
       - [0x30790, .rodata, 3E30C]
       - [0x307B0, .rodata, 3E30C]
       - [0x307C8, rodata]
+      - [0x30834, .rodata, 445BC] # here
       - [0x30858, .rodata, 45568]
       - [0x30880, rodata]
       - [0x3089C, .rodata, 4672C]
       - [0x308B4, rodata]
       - [0x30958, c]
       - [0x3E30C, c]
+      - [0x445BC, c]
       - [0x45568, c]
       - [0x4672C, c]
       - [0x4AA10, sbss]
